requirements.txt
pandas
numpy
matplotlib
seaborn
streamlit
streamlit-toggle-switch
plotly